fix scroll top button animate
