+++ src/css/components/scroll-top.css
@@ -5,25 +5,19 @@
   place-items: center;
   background-color: var(--black);
   border-radius: 50px;
-<<<<<<< Updated upstream
-=======
-
->>>>>>> Stashed changes
   width: 40px;
   height: 40px;
   transition: 0.2s;
   position: fixed;
   right: 20px;
   bottom: 50px;
-
-  box-shadow: inset 0, 0, 0 5px var(--white);
 }
 .scroll-top-icon-wrapper {
   display: grid;
   place-items: center;
   background-color: var(--black);
-  width: 93%;
-  height: 93%;
+  width: 90%;
+  height: 90%;
   border-radius: 50%;
 }
 @media screen and (min-width: 768px) {
@@ -38,15 +32,8 @@
   visibility: hidden;
 }
 .scroll-top-icon {
-<<<<<<< Updated upstream
   max-width: 13px;
   max-height: 13px;
-=======
-  stroke: none;
-  stroke: aqua;
-  width: 13px;
-  height: 13px;
->>>>>>> Stashed changes
   fill: var(--white);
 }
 .js-scroll-top-button:hover .scroll-top-icon-wrapper {
