+++ src/js/handlers/index.js
@@ -13,7 +13,6 @@ import handleFavoritePagination from './favorites-pagination-handler';
 import handlerOpenMenu from './burget-menu-handler';
 import handleFavoritesRemove from './favorites-remove-handler';
 import handlerAddToFavoritesBtn from './add-to-favorites-btn-handler';
-import handleAnimateTopScrollButton from './animate-top-scroll-btn-handler';
 
 export {
   handleFilter,
@@ -32,5 +31,4 @@ export {
   handlerOpenMenu,
   handleFavoritesRemove,
   handlerAddToFavoritesBtn,
-  handleAnimateTopScrollButton,
 };
